+++ app/components/Layout/index.js
@@ -1,7 +1,3 @@
 // @flow
 import Layout from './Layout';
-import Title from './components/Title';
-
 export default Layout;
-
-export { Title };
