+++ app/components/Modals/index.js
@@ -1,0 +1,3 @@
+// @flow
+import Modals from './Modals';
+export default Modals;
