+++ app/components/Sidebar/components/HeaderBlock.js
