+++ app/components/Sidebar/index.js
@@ -1,0 +1,3 @@
+// @flow
+import Sidebar from './Sidebar';
+export default Sidebar;
